Fix Palindrome solution
Fixed a bug where it is assumed that char.GetNumericValue returns the
ASCII value of a char, whereas in reality it actually parses a number
out of a character and returns it as a double. e.g.
Char.GetNumericValue('5') returns 5. Char.GetNumericValue('a') returns
-1. Another bug is that the code only handled lowercase letters, but
does not pre-process the string to convert it to lowercase before
proceeding. The first bug was hiding the second and the net result is
that any string input to the code would return true as a palindrome.
This is also due to lack of error checking, which I did not fix. But at
least now the code returns correct results in non-error cases.